Fixed strings

+++ bbruntime/bbstring.cpp
@@ -24,12 +24,12 @@ BBStr *bbRight( BBStr *s,int n ){
 }
 
 BBStr *bbReplace( BBStr *s,BBStr *from,BBStr *to ){
-	int n=0,from_sz=from->size(),to_sz=to->size();
-	while( n<(int)(s->size() && (n=s->find( *from,n ))!=string::npos) ){
-		s->replace( n,from_sz,*to );
-		n+=to_sz;
+	int n = 0, from_sz = from->size(), to_sz = to->size();
+	while (n < s->size() && (n = s->find(*from, n)) != string::npos) {
+		s->replace(n, from_sz, *to);
+		n += to_sz;
 	}
-	delete from;delete to;return s;
+	delete from; delete to; return s;
 }
 
 int bbInstr( BBStr *s,BBStr *t,int from ){
